Re-organize source, use .hpp extension for C++ headers
