+++ rte/src/brux/core.cpp
@@ -2,9 +2,9 @@
 | CORE SOURCE |
 \*===========*/
 
-#include "main.h"
-#include "global.h"
-#include "core.h"
+#include "brux/main.hpp"
+#include "brux/global.hpp"
+#include "brux/core.hpp"
 
 void xyLoadCore() {
 	const SQChar *cmd =R"rew(
