+++ rte/src/brux/text.hpp
